Returning all systems after Initialize
It seems, like the conversion workflow since .32 without ComponentDataProxy but with IConvertGameObjectToEntity needs all systems returned instead only those which are collected inside DefaultSystems.

Maybe there is an error inside the implementation which systems are all removed for DefaultSystems, which I will have to track later.

This serves as a hotfix to get the new conversion workflow working.